+++ test-update.ps1
@@ -1,3 +1,3 @@
 Push-Location ./test/a
-& cargo run --bin eto -- auto-patch
+& cargo run --bin eto -- auto-patch -p ../../*.etopack
 Pop-Location