+++ test-list.ps1
@@ -1,0 +1,1 @@
+& cargo run --bin eto -- list -p ./package.etopack