+++ test-update.ps1
@@ -1,3 +1,3 @@
 Push-Location ./test/a
-& cargo run --bin eto -- patch -p ../../*.etopack --wait_for 15244 --on_complete taskmgr
+& cargo run --bin eto -- patch -p ../../*.etopack --wait-for 17744 --on-complete notepad
 Pop-Location