+++ changelog.md
@@ -3,6 +3,9 @@
 ## 0.2.0 (upcoming)
 
 - Replace separate eto binaries with sub-commands.
+- Remove "not_running" feature, `eto-updater` (now `eto auto-patch`) should no longer be run
+    directly by a user, an external batch script should check for this instead if you still want
+    this functionality.
 
 ## 0.1.4