+++ changelog.md
@@ -3,12 +3,13 @@
 ## 0.2.0 (upcoming)
 
 - **breaking**: Change package format to JSON + custom GZip + tar data blocks.
+- Add specifying package location to `eto patch`.
+- Add `eto list` subcommand.
+- Add `wait-for` and `on-complete` flags on `patch` subcommand.
 - Replace separate eto binaries with sub-commands.
 - Remove "not_running" feature, `eto-updater` (now `eto patch`) should no longer be run directly by
     a user, an external batch script should check for this instead if you still want this
     functionality.
-- Add specifying package location to `eto patch`.
-- Add `eto list` subcommand.
 
 ## 0.1.4